requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==3.1.12
django-cors-headers==4.6.0
django-mail-templated==2.6.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.7
gunicorn==23.0.0
idna==3.10
packaging==24.2
pipdeptree==2.24.0
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.9.0
pymongo==3.11.4
python-dotenv==1.0.1
pytz==2025.1
razorpay==1.4.2
requests==2.32.3
setuptools==75.6.0
sqlparse==0.2.4
tzdata==2024.2
urllib3==2.2.3
